Cleaned up Classes.
Added Vehicles into the Vehicles.cs file to reduce clutter.
Same for Humans and Resources.
