+++ FireWorks/FireWorks/Human.cs
@@ -17,4 +17,35 @@ namespace FireWorks
         }
         public abstract override string ToString();
     }
+    class FireFighter : Human
+    {
+        public FireFighter(string fname, string lname, int id)
+        {
+            LastName = lname;
+            FirstName = fname;
+            Id = id;
+        }
+        public override string ToString()
+        {
+            return FirstName + " " + LastName + " " + Id;
+        }
+    }
+    class User : Human
+    {
+        public string PIN { get; set; }
+        public string Status { get; set; }
+
+        public User(string fname, string lname, int id, string status, string pin)
+        {
+            LastName = lname;
+            FirstName = fname;
+            Id = id;
+            PIN = pin;
+            Status = status;
+        }
+        public override string ToString()
+        {
+            return FirstName + " " + LastName + " " + Id + " " + PIN + " " + Status;
+        }
+    }
 }
