--- FireWorks/FireWorks/Pkw.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace FireWorks
-{
-    class Pkw : Vehicles
-    {
-
-        public Pkw(string t, int ep, int s) :
-        base(t, ep, s)
-        {
-        }
-    }
-}
