--- FireWorks/FireWorks/TurntableLadder.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace FireWorks
-{
-    class TurntableLadder : ToolCarrier
-    {
-        public int LadderHeight { get; set; }
-        public TurntableLadder(string t, int ep, int s, bool cs, int lh) :
-        base(t, ep, s, cs)
-        {
-            LadderHeight = lh;
-        }
-    }
-}
